+++ qatrack/api/parts/urls.py
@@ -5,7 +5,6 @@ from qatrack.api.parts import views
 
 router = routers.DefaultRouter()
 router.register(r'suppliers', views.SupplierViewSet)
-router.register(r'storages', views.StorageViewSet)
 router.register(r'partcategories', views.PartCategoryViewSet)
 router.register(r'parts', views.PartViewSet)
 router.register(r'partstoragecollections', views.PartStorageCollectionViewSet)