+++ qatrack/api/parts/views.py
@@ -13,13 +13,6 @@ class SupplierViewSet(viewsets.ModelViewSet):
     filter_backends = (backends.RestFrameworkFilterBackend, OrderingFilter,)
 
 
-class StorageViewSet(viewsets.ModelViewSet):
-    queryset = models.Storage.objects.all()
-    serializer_class = serializers.StorageSerializer
-    filterset_class = filters.StorageFilter
-    filter_backends = (backends.RestFrameworkFilterBackend, OrderingFilter,)
-
-
 class PartCategoryViewSet(viewsets.ModelViewSet):
     queryset = models.PartCategory.objects.all()
     serializer_class = serializers.PartCategorySerializer